// data/github/current/ui/packages/commenting/components/issue-comment/__tests__/__generated__/IssueCommentViewerCommentTestQuery.graphql.ts
export type IssueCommentViewerCommentTestQuery$variables = {
  commentId: string;
};
export type IssueCommentViewerCommentTestQuery$data = {
  readonly comment: {
    readonly " $fragmentSpreads": FragmentRefs<"IssueCommentViewerCommentRow" | "IssueCommentViewerReactable">;
  } | null | undefined;
};
export type IssueCommentViewerCommentTestQuery = {
  response: IssueCommentViewerCommentTestQuery$data;
  variables: IssueCommentViewerCommentTestQuery$variables;
};